Move DataManager into storage package
