+++ src/main/java/seedu/duke/command/addtask/TaskCommand.java
@@ -5,7 +5,6 @@ import java.util.Map;
 import seedu.duke.command.Command;
 import seedu.duke.command.CommandResult;
 import seedu.duke.exception.GetTaskFailedException;
-import seedu.duke.local.DataManager;
 import seedu.duke.local.TasktoLineConverter;
 import seedu.duke.parser.TaskUsageParser;
 import seedu.duke.task.Task;
