+++ src/main/java/seedu/duke/task/taskmanager/TaskManager.java
@@ -16,7 +16,7 @@ import seedu.duke.exception.ParseDateFailedException;
 import seedu.duke.exception.SortFormatException;
 import seedu.duke.exception.StartDateAfterEndDateException;
 import seedu.duke.exception.TaskIsNonRecurringException;
-import seedu.duke.local.DataManager;
+import seedu.duke.storage.DataManager;
 import seedu.duke.log.Log;
 
 import java.time.LocalDateTime;
